Display the base R colour options.

display the base R colour options in the active graphics plotting device or printed to a PDF file. This makes it easier to see what the colour options are in conjunction with their R labels to make it easier to specify custom colours for plots.

colour_options(print_to_pdf = FALSE, pdf_name = "base_r_colour_options.pdf")

Arguments

print_to_pdf
IF TRUE, output is exported to a PDF in the working directory instead of displayed in an R graphics device.

pdf_name
The name of the PDF file. Default is "base_r_colour_options.pdf". Must end in .pdf. Ignored if print_to_pdf = FALSE.

References

Code was adapted from http://bc.bojanorama.pl/2013/04/r-color-reference-sheet/

Author

Craig P. Hutton, Craig.Hutton@gov.bc.ca

Examples

if (FALSE) {
#default settings. Prints to the active graphics device.
colour_options(print_to_pdf = FALSE) #the default
colour_options(print_to_pdf = TRUE) #print to a single page PDF file in your working directory.
}