practice data for package testing and development

A dataset containing randomly generated practice data of various classes for testing and developing elucidate package functions.

pdata

Format

A data frame of 12,000 rows and 10 columns

id

a sequence of numbers from 1:1000 labelling the row id, repeats for each unique value of d

d

a sequence of dates (yyyy-mm-dd) ranging from 2008-01-01 to 2019-01-01 in 1 year increments

g

5 level factor of values "a", "b", "c", "d", and "e" which was assigned to ids randomly within d

high_low

character vector indicating whether y2 is above = "high" or below = "low" the mean value of 100

even

logical vector indicating whether a row id is even = TRUE or odd = FALSE

y1

random samples from a normal distribution within each level of g and d

y2

random sample from a normal distribution with mean of 100 and standard deviation of 10

x1

random sample of values between 1-100

x2

random sample of values between 101-200

x3

random sample of values between 201-300

Source

"Randomly generated data"